% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITKR2-package.R
\name{itk_info}
\alias{itk_info}
\title{Get 'ITK' information}
\usage{
itk_info()
}
\value{
A list of following items:
\describe{
\item{\code{version}}{numerical version}
\item{\code{release}}{release version}
\item{\code{include}}{path to header folders}
\item{\code{lib}}{path to library folder}
}
}
\description{
'ITK' versions, compile information
}
\examples{

itk_info()


}
